fix: properly read namespaces with no prefix from the root element
Another attempt to fix: https://github.com/crewjam/saml/pull/580

The code is reading the attributes from the elements in the loop,
then writes to a map using `chileEl.Scope` as the key.
Intention behind this code is not clear, as the map contains a random
string for an empty space, depending which element was the last one.

Instead, write the key to the map only if the `childEl` is the element
we're encoding.

Tested that against Fusion Auth SAML and Microsoft SP with SAML.
Can't submit the ACS/Metadata for Microsoft for the testing purposes, as
it contains the sensitive data.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>